Use base64 encoding for protobuf in grid
