+++ GLOCKFILE
@@ -15,10 +15,10 @@ github.com/kr/pretty e6ac2fc51e89a3249e82157fa0bb7a18ef9dd5bb
 github.com/kr/text bb797dc4fb8320488f47bf11de07a733d7233e1f
 github.com/leekchan/timeutil 28917288c48df3d2c1cfe468c273e0b2adda0aa5
 github.com/lytics/cloudstorage 6e53bda60adbcbdb8b9184bef1fbd1fd2f6aae3d
-github.com/lytics/confl fc7bee8fab5c930d8947e22c9da6fd6948b6e3a7
+github.com/lytics/confl f12b6bace2ff295fc6730f6dbaf009289e2f6e80
 github.com/lytics/datemath 988020f3ad34814005ab10b6c7863e31672b5f63
 github.com/lytics/dfa 63e35f788f7fa5203fcd2dcd7e318da5a1b981e5
-github.com/lytics/grid 44362b9bb11364b93840fa7ce704d4099c27392a
+github.com/lytics/grid 892123196528d64b4633dfd8c91d6aee8d69d22d
 github.com/lytics/metafora 8ec2f856db6abe826989b255a02908fc3f0c5eef
 github.com/lytics/sereno c1ca998b0b3a7b2127ee9ac844753a430a1aecae
 github.com/mattbaird/elastigo 6846107dcb6b61f78681686b1d6b27787aa8f881
