+++ dataux.conf
@@ -1,5 +1,5 @@
 # 
-# dataux configuration
+# dataux configuration example
 #
 
 supress_recover: true
@@ -16,8 +16,8 @@ frontends [
   }
 ]
 
-# schemas:   ie the virtual database made up of 
-#   combining tables from each source into a flattened table pool
+# schemas:   ie a virtual database made up of 
+#   combining tables from each source into a flattened table
 schemas : [
   {
     name : datauxtest
@@ -25,12 +25,14 @@ schemas : [
   },
   {
     name : baseball
-    sources : [ "gcscsvs2" ]
+    sources : [ "gcs_baseball_csv" ]
   }
 ]
 
 # sources
-#   Note the google-datastore requires JWT Tokens
+#   - elasticsearch which uses example data from github, see tools/importgithub
+#   - cloud files w baseball csv files
+#   - mongo
 #
 sources : [
   {
@@ -42,15 +44,16 @@ sources : [
     type : elasticsearch
   },
   {
-    # this section is for csv files from
-    #    http://seanlahman.com/baseball-archive/statistics/
-    name     : "gcscsvs2"
-    type     : "cloudstore"
-    settings : {
-      type             : "gcs"
-      gcsbucket        : "lytics-dataux-tests"
-      path             : "baseball/"
-      format           : "csv"
+    # this section is a "CloudFiles" source for csv files 
+    # stored on google-storage from
+    #      http://seanlahman.com/baseball-archive/statistics/ 
+    name     "gcs_baseball_csv"
+    type     "cloudstore"
+    settings  {
+      type           "gcs"                   # "gcs","localfs", ??
+      gcsbucket      "lytics-dataux-tests"
+      path           "baseball/"
+      format         "csv"
     }
   },
   {
